0.4.1

+++ src/main/java/openwrestling/model/gameObjects/Contract.java
@@ -18,24 +18,16 @@ import java.time.LocalDate;
 @AllArgsConstructor
 public class Contract extends GameObject implements Serializable, iContract {
 
-    private int contractID;
+    private long contractID;
     private Promotion promotion;
     private Worker worker;
-
     private boolean active = true;
-
     private LocalDate startDate;
-
     private LocalDate endDate;
-
     private LocalDate lastShowDate;
-
     private boolean exclusive;
-
     private boolean pushed;
-
     private int appearanceCost;
-
     private int monthlyCost;
 
     public Contract(LocalDate startDate, Worker worker, Promotion promotion) {
@@ -58,7 +50,7 @@ public class Contract extends GameObject implements Serializable, iContract {
     }
 
     public LocalDate getLastShowDate() {
-        if(lastShowDate == null) {
+        if (lastShowDate == null) {
             return startDate;
         }
         return lastShowDate;
